untangle: listener rebind utils descend to intendant-core::net
should_continue_after_accept_error / rebind_dead_tcp_listener /
FATAL_ACCEPT_REBIND_THRESHOLD are generic TCP-listener lifecycle
policy shared by the dashboard gateway, the control socket, and the
enrollment cert server -- the last one an access -> web_gateway ref,
the wrong direction for the layering. They join core::net beside the
existing probes, tests (incl. the AddrInUse-patience pair) travel
along, and listener.rs re-exports at the old mount so the gateway and
control-socket call sites compile untouched. Core gains tokio net +
socket2; dev tokio grows macros/rt/time for the accept-loop tests.
