+++ crates/intendant-core/Cargo.toml
@@ -11,7 +11,8 @@ serde = { version = "1.0", features = ["derive"] }
 # skills discovery walks the home skills dir.
 dirs = "6"
 serde_json = "1.0"
-tokio = { version = "1.0", features = ["sync"] }
+socket2 = "0.6"
+tokio = { version = "1.0", features = ["net", "sync"] }
 # CallerError carries reqwest::Error in its Http variant; feature set
 # mirrors the caller's so workspace feature unification stays a no-op.
 reqwest = { version = "0.12", features = ["json", "rustls-tls", "stream", "multipart"], default-features = false }
@@ -21,6 +22,8 @@ reqwest = { version = "0.12", features = ["json", "rustls-tls", "stream", "multi
 # write registries into temp dirs.
 toml = "0.8"
 tempfile = "3"
+# the listener rebind tests drive a real accept loop
+tokio = { version = "1.0", features = ["macros", "net", "rt-multi-thread", "time"] }
 
 [target.'cfg(unix)'.dependencies]
 libc = "0.2"
